Phase 1: resolve import cycle and export the node-model public API
Restore the node_table/collect lint fixes (break the basedpyright import cycle)
and export the new public symbols from chopdiff.docs: Node, NodeKind, Layer,
NestingGuarantee, LAYER_NESTING, NodeTable, build_node_table, SpanRef,
base_blocks, BaseBlock, collect.

Bead chopdiff-1ehf (integration).
