+++ src/chopdiff/docs/__init__.py
@@ -1,9 +1,14 @@
 # flake8: noqa: F401
 
+from chopdiff.docs.base_blocks import BaseBlock, base_blocks
 from chopdiff.docs.block_tree import Block
 from chopdiff.docs.block_types import BlockType
+from chopdiff.docs.collect import collect
+from chopdiff.docs.node import LAYER_NESTING, Layer, NestingGuarantee, Node, NodeKind, NodeTable
+from chopdiff.docs.node_table import build_node_table
 from chopdiff.docs.search_tokens import search_tokens
 from chopdiff.docs.sizes import TextUnit
+from chopdiff.docs.span_ref import SpanRef
 from chopdiff.docs.text_doc import (
     Link,
     Offsets,
@@ -102,4 +107,15 @@ __all__ = [
     "wordtok_to_str",
     "wordtokenize",
     "wordtokenize_with_offsets",
+    "BaseBlock",
+    "base_blocks",
+    "collect",
+    "LAYER_NESTING",
+    "Layer",
+    "NestingGuarantee",
+    "Node",
+    "NodeKind",
+    "NodeTable",
+    "build_node_table",
+    "SpanRef",
 ]
